move test around
