+++ sklearn/tests/test_common.py
@@ -14,7 +14,6 @@ import pkgutil
 from inspect import isgenerator
 from functools import partial
 
-import numpy as np
 import pytest
 
 from sklearn.utils import all_estimators
@@ -42,12 +41,6 @@ from sklearn.utils.estimator_checks import (
     check_class_weight_balanced_linear_classifier,
     parametrize_with_checks,
     check_n_features_in_after_fitting,
-)
-from sklearn.multiclass import check_classification_targets
-from sklearn.utils.validation import (
-    check_array,
-    check_is_fitted,
-    check_X_y,
 )
 
 
@@ -304,186 +297,3 @@ N_FEATURES_IN_AFTER_FIT_ESTIMATORS = [
 def test_check_n_features_in_after_fitting(estimator):
     _set_checking_parameters(estimator)
     check_n_features_in_after_fitting(estimator.__class__.__name__, estimator)
-
-
-class MinimalClassifier:
-    """Minimal classifier implementation with inheriting from BaseEstimator."""
-    _estimator_type = "classifier"
-
-    def __init__(self, param=None):
-        self.param = param
-
-    def __repr__(self):
-        # Only required when using pytest-xdist to get an id not associated
-        # with the memory location
-        return self.__class__.__name__
-
-    def get_params(self, deep=True):
-        return {"param": self.param}
-
-    def set_params(self, **params):
-        for key, value in params.items():
-            setattr(self, key, value)
-        return self
-
-    def __getstate__(self):
-        return self.__dict__.copy()
-
-    def __setstate__(self, state):
-        self.__dict__.update(state)
-
-    def fit(self, X, y):
-        X, y = check_X_y(X, y)
-        check_classification_targets(y)
-        self.n_features_in_ = X.shape[1]
-        self.classes_, counts = np.unique(y, return_counts=True)
-        self._most_frequent_class_idx = counts.argmax()
-        return self
-
-    def predict_proba(self, X):
-        check_is_fitted(self)
-        X = check_array(X)
-        if X.shape[1] != self.n_features_in_:
-            raise ValueError
-        proba_shape = (X.shape[0], self.classes_.size)
-        y_proba = np.zeros(shape=proba_shape, dtype=np.float64)
-        y_proba[:, self._most_frequent_class_idx] = 1.0
-        return y_proba
-
-    def predict(self, X):
-        y_proba = self.predict_proba(X)
-        y_pred = y_proba.argmax(axis=1)
-        return self.classes_[y_pred]
-
-    def score(self, X, y):
-        from sklearn.metrics import accuracy_score
-        return accuracy_score(y, self.predict(X))
-
-
-class MinimalRegressor:
-    """Minimal regressor implementation with inheriting from BaseEstimator."""
-    _estimator_type = "regressor"
-
-    def __init__(self, param=None):
-        self.param = param
-
-    def __repr__(self):
-        # Only required when using pytest-xdist to get an id not associated
-        # with the memory location
-        return self.__class__.__name__
-
-    def get_params(self, deep=True):
-        return {"param": self.param}
-
-    def set_params(self, **params):
-        for key, value in params.items():
-            setattr(self, key, value)
-        return self
-
-    def __getstate__(self):
-        return self.__dict__.copy()
-
-    def __setstate__(self, state):
-        self.__dict__.update(state)
-
-    def fit(self, X, y):
-        X, y = check_X_y(X, y)
-        self.n_features_in_ = X.shape[1]
-        self._mean = np.mean(y)
-        return self
-
-    def predict(self, X):
-        check_is_fitted(self)
-        X = check_array(X)
-        if X.shape[1] != self.n_features_in_:
-            raise ValueError
-        return np.ones(shape=(X.shape[0],)) * self._mean
-
-    def score(self, X, y):
-        from sklearn.metrics import r2_score
-        return r2_score(y, self.predict(X))
-
-
-class MinimalTransformer:
-    """Minimal transformer implementation with inheriting from
-    BaseEstimator."""
-
-    def __init__(self, param=None):
-        self.param = param
-
-    def __repr__(self):
-        # Only required when using pytest-xdist to get an id not associated
-        # with the memory location
-        return self.__class__.__name__
-
-    def get_params(self, deep=True):
-        return {"param": self.param}
-
-    def set_params(self, **params):
-        for key, value in params.items():
-            setattr(self, key, value)
-        return self
-
-    def __getstate__(self):
-        return self.__dict__.copy()
-
-    def __setstate__(self, state):
-        self.__dict__.update(state)
-
-    def fit(self, X, y=None):
-        X = check_array(X)
-        self.n_features_in_ = X.shape[1]
-        return self
-
-    def transform(self, X, y=None):
-        check_is_fitted(self)
-        X = check_array(X)
-        if X.shape[1] != self.n_features_in_:
-            raise ValueError
-        return X
-
-    def fit_transform(self, X, y=None):
-        return self.fit(X, y).transform(X, y)
-
-
-def _generate_minimal_compatible_instances():
-    """Generate instance containing estimators from minimal class compatible
-    implementation."""
-    for SearchCV, (Estimator, param_grid) in zip(
-        [GridSearchCV, RandomizedSearchCV],
-        [
-            (MinimalRegressor, {"param": [1, 10]}),
-            (MinimalClassifier, {"param": [1, 10]}),
-        ],
-    ):
-        yield SearchCV(Estimator(), param_grid)
-
-    for SearchCV, (Estimator, param_grid) in zip(
-        [GridSearchCV, RandomizedSearchCV],
-        [
-            (MinimalRegressor, {
-                "minimaltransformer__param": [1, 10],
-                "minimalregressor__param": [1, 10],
-            }),
-            (MinimalClassifier, {
-                "minimaltransformer__param": [1, 10],
-                "minimalclassifier__param": [1, 10],
-            }),
-        ],
-    ):
-        yield SearchCV(
-            make_pipeline(MinimalTransformer(), Estimator()), param_grid
-        ).set_params(error_score="raise")
-
-
-# FIXME: hopefully in 0.25
-@pytest.mark.xfail(
-    reason=("This test is currently failing because checks are granular "
-            "enough. Once checks are split with some kind of only API tests, "
-            "this test should enabled.")
-)
-@parametrize_with_checks(list(_generate_minimal_compatible_instances()))
-def test_minimal_class_implementation_checks(estimator, check):
-    # Check that third-party library can run tests without inheriting from
-    # BaseEstimator.
-    check(estimator)
